refactor: Menu, MenuProduct 를 JPA 엔티티로 변경, Repository 적용
