+++ src/main/java/kitchenpos/dao/JdbcTemplateMenuDao.java
@@ -3,9 +3,10 @@ package kitchenpos.dao;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import javax.sql.DataSource;
-import kitchenpos.domain.Menu;
+import kitchenpos.domain.menu.Menu;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
@@ -16,6 +17,7 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public class JdbcTemplateMenuDao implements MenuDao {
+
     private static final String TABLE_NAME = "menu";
     private static final String KEY_COLUMN_NAME = "id";
 
@@ -68,11 +70,9 @@ public class JdbcTemplateMenuDao implements MenuDao {
     }
 
     private Menu toEntity(final ResultSet resultSet) throws SQLException {
-        final Menu entity = new Menu();
-        entity.setId(resultSet.getLong("id"));
-        entity.setName(resultSet.getString("name"));
-        entity.setPrice(resultSet.getBigDecimal("price"));
-        entity.setMenuGroupId(resultSet.getLong("menu_group_id"));
+        final Menu entity = Menu.of(resultSet.getLong("id"), resultSet.getString("name"),
+                resultSet.getBigDecimal("price"), resultSet.getLong("menu_group_id"), Map.of());
         return entity;
     }
+
 }
