+++ src/main/java/kitchenpos/dao/MenuDao.java
@@ -2,7 +2,7 @@ package kitchenpos.dao;
 
 import java.util.List;
 import java.util.Optional;
-import kitchenpos.domain.Menu;
+import kitchenpos.domain.menu.Menu;
 
 public interface MenuDao {
     Menu save(Menu entity);
