+++ src/main/java/kitchenpos/domain/menu/Menu.java
@@ -1,0 +1,96 @@
+package kitchenpos.domain.menu;
+
+import java.math.BigDecimal;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.stream.Collectors;
+import javax.persistence.CascadeType;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.ManyToOne;
+import javax.persistence.OneToMany;
+import kitchenpos.domain.Product;
+import kitchenpos.domain.menugroup.MenuGroup;
+
+import static javax.persistence.CascadeType.PERSIST;
+import static javax.persistence.CascadeType.REMOVE;
+
+@Entity
+public class Menu {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String name;
+    private BigDecimal price;
+    private Long menuGroupId;
+
+    @OneToMany(mappedBy = "menu", cascade = {PERSIST, REMOVE}, orphanRemoval = true)
+    private List<MenuProduct> menuProducts = new ArrayList<>();
+
+    protected Menu() {
+    }
+
+    public static Menu of(final Long id, final String name, final BigDecimal price, final Long menuGroupId,
+                          final Map<Product, Long> productToQuantity) {
+        Menu menu = new Menu(id, name, price, menuGroupId);
+        menu.changeMenuProducts(productToQuantity);
+        return menu;
+    }
+
+    private Menu(final Long id, final String name, final BigDecimal price, final Long menuGroupId) {
+        if (Objects.isNull(price) || price.compareTo(BigDecimal.ZERO) < 0) {
+            throw new IllegalArgumentException("메뉴 가격이 없거나 0보다 작습니다.");
+        }
+        this.id = id;
+        this.name = name;
+        this.price = price;
+        this.menuGroupId = menuGroupId;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(final Long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(final String name) {
+        this.name = name;
+    }
+
+    public BigDecimal getPrice() {
+        return price;
+    }
+
+    public void setPrice(final BigDecimal price) {
+        this.price = price;
+    }
+    public List<MenuProduct> getMenuProducts() {
+        return menuProducts;
+    }
+
+    public void changeMenuProducts(final Map<Product, Long> productToQuantity) {
+        final var sumOfMenuProductPrice = productToQuantity.entrySet().stream()
+                .map(entry -> entry.getKey().getPrice().multiply(BigDecimal.valueOf(entry.getValue())))
+                .reduce(BigDecimal.ZERO, BigDecimal::add);
+
+        if (price.compareTo(sumOfMenuProductPrice) == 1) {
+            throw new IllegalArgumentException("메뉴 가격이 메뉴에 속한 상품 가격의 합보다 큽니다.");
+        }
+        final List<MenuProduct> menuProducts = productToQuantity.entrySet().stream()
+                .map(entry -> new MenuProduct(null, this, entry.getKey().getId(), entry.getValue()))
+                .collect(Collectors.toList());
+        this.menuProducts = menuProducts;
+    }
+
+}
