+++ src/main/java/kitchenpos/domain/menu/MenuProduct.java
@@ -1,0 +1,37 @@
+package kitchenpos.domain.menu;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.ManyToOne;
+
+@Entity
+public class MenuProduct {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long seq;
+
+    @ManyToOne
+    private Menu menu;
+
+    private Long productId;
+
+    private long quantity;
+
+    protected MenuProduct() {
+    }
+
+    public MenuProduct(final Long seq, final Menu menu, final Long productId, final long quantity) {
+        this.seq = seq;
+        this.menu = menu;
+        this.productId = productId;
+        this.quantity = quantity;
+    }
+
+    public void setMenu(final Menu menu) {
+        this.menu = menu;
+    }
+
+}
