+++ src/main/java/kitchenpos/domain/menu/MenuRepository.java
@@ -1,0 +1,10 @@
+package kitchenpos.domain.menu;
+
+import java.util.List;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MenuRepository extends JpaRepository<Menu, Long> {
+
+    long countByIdIn(List<Long> ids);
+
+}
