+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -113,7 +113,7 @@ class OrderServiceTest extends ServiceTest {
     }
 
     private OrderLineItemRequest 주문_상품_초기화() {
-        final var 메뉴 = menuDao.findAll().get(0);
+        final var 메뉴 = menuRepository.findAll().get(0);
         return new OrderLineItemRequest(메뉴.getId(), 1);
     }
