+++ src/test/java/kitchenpos/application/ServiceTest.java
@@ -1,11 +1,10 @@
 package kitchenpos.application;
 
-import kitchenpos.dao.MenuDao;
 import kitchenpos.dao.MenuGroupDao;
-import kitchenpos.dao.MenuProductDao;
 import kitchenpos.dao.OrderTableDao;
 import kitchenpos.dao.ProductDao;
 import kitchenpos.dao.TableGroupDao;
+import kitchenpos.domain.menu.MenuRepository;
 import kitchenpos.domain.order.OrderRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -16,13 +15,10 @@ import org.springframework.transaction.annotation.Transactional;
 public abstract class ServiceTest {
 
     @Autowired
+    protected MenuRepository menuRepository;
+
+    @Autowired
     protected ProductDao productDao;
-
-    @Autowired
-    protected MenuProductDao menuProductDao;
-
-    @Autowired
-    protected MenuDao menuDao;
 
     @Autowired
     protected MenuGroupDao menuGroupDao;
