+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -162,7 +162,7 @@ class TableServiceTest extends ServiceTest {
         }
 
         private OrderLineItem 주문_상품_만들기() {
-            final var 메뉴 = menuDao.findAll().get(0);
+            final var 메뉴 = menuRepository.findAll().get(0);
             return new OrderLineItem(null, null, 메뉴.getId(), 1);
         }
